perf: 可指定版本

+++ tools/sync-reset.sh
@@ -15,10 +15,14 @@ root_dir=$(cd `dirname $0`/.. && pwd -P)
 
 source_dir="$root_dir/source-code"
 nwjs_dir="$source_dir/nwjs"
-dirs=("src" "src/third_party/boringssl/src" "src/third_party/devtools-frontend/src" "src/third_party/ffmpeg" "src/third_party/icu" "src/third_party/lss" "src/third_party/webdriver/pylib" "src/third_party/swiftshader" "src/tools/gyp")
+dirs=("src" "src/third_party/boringssl/src" "src/third_party/devtools-frontend/src" "src/third_party/ffmpeg" "src/third_party/icu" "src/third_party/lss" "src/third_party/webdriver/pylib" "src/third_party/swiftshader" "src/tools/gyp" "src/third_party/perfetto" "src/third_party/vulkan-deps/vulkan-validation-layers/src" "src/third_party/breakpad/breakpad" "src/third_party/dawn" "src/third_party/webrtc")
 for dir in ${dirs[@]};
 do
   target_dir="$nwjs_dir/$dir"
-  notice "clean: $target_dir"
-  cd $target_dir && git checkout . && git clean -xdf
+  if [ -d "$target_dir" ];then
+    notice "clean: $target_dir"
+    cd $target_dir && git checkout . && git clean -xdf
+  else
+    notice "skip: $target_dir"
+  fi
 done
